+++ tests/Fixtures/ComplexEntity.php
@@ -1,24 +1,18 @@
 <?php declare(strict_types=1);
 
-namespace MF\Tests\Fixtures;
+namespace MF\Collection\Fixtures;
 
 class ComplexEntity implements EntityInterface
 {
     /** @var SimpleEntity */
     private $simpleEntity;
 
-    /**
-     * @param SimpleEntity $simpleEntity
-     */
     public function __construct(SimpleEntity $simpleEntity)
     {
         $this->simpleEntity = $simpleEntity;
     }
 
-    /**
-     * @return SimpleEntity
-     */
-    public function getSimpleEntity()
+    public function getSimpleEntity(): SimpleEntity
     {
         return $this->simpleEntity;
     }