+++ tests/Immutable/Generic/ListTest.php
@@ -1,21 +1,20 @@
 <?php declare(strict_types=1);
 
-namespace MF\Tests\Collection\Immutable\Generic;
+namespace MF\Collection\Immutable\Generic;
 
 use Eris\Generator;
+use MF\Collection\AbstractTestCase;
+use MF\Collection\Fixtures\ComplexEntity;
+use MF\Collection\Fixtures\EntityInterface;
+use MF\Collection\Fixtures\SimpleEntity;
 use MF\Collection\Generic\ICollection;
 use MF\Collection\Generic\IList as GenericListInterface;
 use MF\Collection\ICollection as BaseCollectionInterface;
 use MF\Collection\IList as BaseListInterface;
 use MF\Collection\Immutable\Generic\IList as ImmutableGenericInterface;
-use MF\Collection\Immutable\Generic\ListCollection;
 use MF\Collection\Immutable\IList;
 use MF\Collection\Immutable\ListCollection as BaseImmutableListCollection;
 use MF\Collection\Mutable\ListCollection as MutableListCollection;
-use MF\Tests\AbstractTestCase;
-use MF\Tests\Fixtures\ComplexEntity;
-use MF\Tests\Fixtures\EntityInterface;
-use MF\Tests\Fixtures\SimpleEntity;
 use MF\Validator\TypeValidator;
 
 class ListTest extends AbstractTestCase