+++ tests/Immutable/MapTest.php
@@ -1,13 +1,9 @@
 <?php declare(strict_types=1);
 
-namespace MF\Tests\Collection\Mutable;
+namespace MF\Collection\Immutable;
 
-use MF\Collection\IMap as BaseMapInterface;
-use MF\Collection\Mutable\ICollection;
-use MF\Collection\Mutable\IList;
-use MF\Collection\Mutable\IMap;
-use MF\Collection\Mutable\Map;
-use MF\Tests\AbstractTestCase;
+use MF\Collection\AbstractTestCase;
+use MF\Collection\ICollection;
 
 class MapTest extends AbstractTestCase
 {
@@ -21,7 +17,6 @@ class MapTest extends AbstractTestCase
 
     public function testShouldImplementsInterfaces(): void
     {
-        $this->assertInstanceOf(BaseMapInterface::class, $this->map);
         $this->assertInstanceOf(IMap::class, $this->map);
         $this->assertInstanceOf(ICollection::class, $this->map);
         $this->assertInstanceOf(\ArrayAccess::class, $this->map);
@@ -112,17 +107,11 @@ class MapTest extends AbstractTestCase
         $this->assertSame([1, 2, 3], $map->toArray());
     }
 
-    /**
-     * @param mixed $key
-     * @param mixed $value
-     *
-     * @dataProvider addItemsProvider
-     */
-    public function testShouldAddItemsToMapArrayWay($key, $value): void
+    public function testShouldThrowBadMethodCallExceptionOnAddItemsToMapArrayWay(): void
     {
-        $this->map[$key] = $value;
+        $this->expectException(\BadMethodCallException::class);
 
-        $this->assertEquals($value, $this->map[$key]);
+        $this->map['key'] = 'value';
     }
 
     /**
@@ -133,9 +122,10 @@ class MapTest extends AbstractTestCase
      */
     public function testShouldAddItemsToMap($key, $value): void
     {
-        $this->map->set($key, $value);
+        $newMap = $this->map->set($key, $value);
 
-        $this->assertEquals($value, $this->map->get($key));
+        $this->assertNotSame($this->map, $newMap);
+        $this->assertEquals($value, $newMap->get($key));
     }
 
     public function addItemsProvider()
@@ -181,9 +171,9 @@ class MapTest extends AbstractTestCase
      *
      * @dataProvider invalidKeyProvider
      */
-    public function testShouldThrowInvalidArgumentExceptionOnAddingObject($key): void
+    public function testShouldThrowBadMethodCallExceptionOnAddingObject($key): void
     {
-        $this->expectException(\InvalidArgumentException::class);
+        $this->expectException(\BadMethodCallException::class);
 
         $this->map[$key] = 'value';
     }
@@ -228,14 +218,9 @@ class MapTest extends AbstractTestCase
 
         $this->assertCount($originalCount, $map);
 
-        $map->set('key', 'value');
-        $this->assertCount($originalCount + 1, $map);
-
-        $map['key'] = 'value X';
-        $this->assertCount($originalCount + 1, $map);
-
-        $map['keyY'] = 'value Y';
-        $this->assertCount($originalCount + 2, $map);
+        $newMap = $map->set('key', 'value');
+        $this->assertCount($originalCount, $map);
+        $this->assertCount($originalCount + 1, $newMap);
     }
 
     public function testShouldHasKeys(): void
@@ -243,7 +228,7 @@ class MapTest extends AbstractTestCase
         $keyExists = 'has-key';
         $keyDoesNotExist = 'has-no-key';
 
-        $this->map->set($keyExists, 'value');
+        $this->map = $this->map->set($keyExists, 'value');
 
         $this->assertArrayHasKey($keyExists, $this->map);
         $this->assertArrayNotHasKey($keyDoesNotExist, $this->map);
@@ -257,17 +242,24 @@ class MapTest extends AbstractTestCase
         $key = 'key';
         $key2 = 'key2';
 
-        $this->map->set($key, 'value');
+        $this->map = $this->map->set($key, 'value');
         $this->assertTrue($this->map->containsKey($key));
 
-        $this->map[$key2] = 'value2';
+        $this->map = $this->map->set($key2, 'value2');
         $this->assertTrue($this->map->containsKey($key2));
 
-        $this->map->remove($key);
-        $this->assertFalse($this->map->containsKey($key));
+        $newMap = $this->map->remove($key);
+        $this->assertTrue($this->map->containsKey($key));
+        $this->assertFalse($newMap->containsKey($key));
+    }
 
-        unset($this->map[$key2]);
-        $this->assertFalse($this->map->containsKey($key2));
+    public function testShouldThrowBadMethodCallExceptionOnUnsetValueArrayWay(): void
+    {
+        $this->expectException(\BadMethodCallException::class);
+
+        $this->map = $this->map->set('key', 'value');
+
+        unset($this->map['key']);
     }
 
     public function testShouldContainsValue(): void
@@ -276,7 +268,7 @@ class MapTest extends AbstractTestCase
         $value = 1;
         $valueNotPresented = 4;
 
-        $this->map->set($key, $value);
+        $this->map = $this->map->set($key, $value);
 
         $this->assertTrue($this->map->contains($value));
         $this->assertFalse($this->map->contains($valueNotPresented));
@@ -348,7 +340,7 @@ class MapTest extends AbstractTestCase
 
     public function testShouldGetValueArrayWay(): void
     {
-        $this->map->set('key', 'value');
+        $this->map = $this->map->set('key', 'value');
 
         $this->assertEquals('value', $this->map['key']);
         $this->assertEquals('value', $this->map->get('key'));
@@ -366,7 +358,7 @@ class MapTest extends AbstractTestCase
 
     public function testShouldCallReducerCorrectly(): void
     {
-        $this->map->set('key', 'value');
+        $this->map = $this->map->set('key', 'value');
 
         $reduced = $this->map->reduce(function ($total, $current, $key, $map) {
             $this->assertEquals('initial', $total);
@@ -390,7 +382,7 @@ class MapTest extends AbstractTestCase
     public function testShouldReduceMap(callable $reducer, array $values, $expected): void
     {
         foreach ($values as $key => $value) {
-            $this->map->set($key, $value);
+            $this->map = $this->map->set($key, $value);
         }
 
         $this->assertEquals($expected, $this->map->reduce($reducer));
@@ -429,7 +421,7 @@ class MapTest extends AbstractTestCase
     public function testShouldReduceMapWithInitialValue(callable $reducer, array $values, $initialValue, $expected): void
     {
         foreach ($values as $key => $value) {
-            $this->map->set($key, $value);
+            $this->map = $this->map->set($key, $value);
         }
 
         $this->assertEquals($expected, $this->map->reduce($reducer, $initialValue));
@@ -467,77 +459,33 @@ class MapTest extends AbstractTestCase
         ];
     }
 
-    public function testShouldGetMutableListAsImmutable(): void
+    public function testShouldGetImmutableMapAsMutable(): void
     {
-        $this->map->set('key', 'value');
+        $this->map = $this->map->set('key', 'value');
 
-        $immutable = $this->map->asImmutable();
+        $mutable = $this->map->asMutable();
 
-        $this->assertInstanceOf(\MF\Collection\Immutable\IMap::class, $immutable);
-        $this->assertInstanceOf(\MF\Collection\Immutable\Map::class, $immutable);
+        $this->assertInstanceOf(\MF\Collection\IMap::class, $mutable);
+        $this->assertInstanceOf(\MF\Collection\Mutable\Map::class, $mutable);
 
-        $this->assertEquals($this->map->toArray(), $immutable->toArray());
+        $this->assertEquals($this->map->toArray(), $mutable->toArray());
     }
 
     public function testShouldClearCollection(): void
     {
-        $this->map->set('key', 'value');
+        $this->map = $this->map->set('key', 'value');
         $this->assertTrue($this->map->contains('value'));
 
-        $this->map->clear();
+        $this->map = $this->map->clear();
         $this->assertFalse($this->map->contains('value'));
     }
 
     public function testShouldCheckIfCollectionIsEmpty(): void
     {
-        $this->map->set('key', 'value');
+        $this->map = $this->map->set('key', 'value');
         $this->assertFalse($this->map->isEmpty());
 
-        $this->map->clear();
+        $this->map = $this->map->clear();
         $this->assertTrue($this->map->isEmpty());
-    }
-
-    public function testShouldMapBigCollectionManyTimesInOneLoop(): void
-    {
-        $this->startTimer();
-        $bigMap = Map::from(range(0, 10000));
-        $creatingCollection = $this->stopTimer();
-
-        $this->startTimer();
-        foreach ($bigMap as $i) {
-            $this->ignore($i);
-        }
-        $loopTime = $this->stopTimer();
-
-        $this->startTimer();
-        $bigMap
-            ->map(function ($v) {
-                return $v + 1;
-            })
-            ->map(function ($v) {
-                return $v * 2;
-            })
-            ->filter(function ($v) {
-                return $v % 2 === 0;
-            })
-            ->map(function ($v) {
-                return $v - 1;
-            });
-        $mappingTime = $this->stopTimer();
-
-        $this->startTimer();
-        foreach ($bigMap as $i) {
-            $this->ignore($i);
-        }
-        $loopWithMappingTime = $this->stopTimer();
-
-        $totalTime = $creatingCollection + $loopTime + $mappingTime + $loopWithMappingTime;
-
-        $this->assertLessThan(1, $mappingTime);
-        $this->assertLessThan($loopTime * 1.5, $loopWithMappingTime);   // 50% is still fair enough
-        $this->assertCount(10001, $bigMap);
-
-        // this test before lazy mapping lasts around 5-6 seconds, and now it is less than 2 seconds
-        $this->assertLessThan(2000, $totalTime);
     }
 }